+++ cmd/juno/main.go
@@ -6,14 +6,14 @@ import (
 	"os/signal"
 	"syscall"
 
-	"github.com/NethermindEth/juno/internal/juno"
+	"github.com/NethermindEth/juno/internal/node"
 )
 
 func main() {
 	quit := make(chan os.Signal, 1)
 	signal.Notify(quit, os.Interrupt, syscall.SIGTERM)
 
-	if err := NewCmd(juno.New, quit).Execute(); err != nil {
+	if err := NewCmd(node.New, quit).Execute(); err != nil {
 		fmt.Fprintln(os.Stderr, err.Error())
 		os.Exit(1)
 	}
